cx_utils: rename lock_shared_ptr to snake case for consistancy
