+++ cx_plugins/ament_index_plugin/include/cx_ament_index_plugin/ament_index_plugin.hpp
@@ -6,7 +6,7 @@
 #include <string>
 
 #include "cx_plugin/clips_plugin.hpp"
-#include "cx_utils/LockSharedPtr.hpp"
+#include "cx_utils/lock_shared_ptr.hpp"
 
 namespace cx {
