+++ cx_plugins/executive_plugin/include/cx_executive_plugin/executive_plugin.hpp
@@ -6,7 +6,7 @@
 #include <string>
 
 #include "cx_plugin/clips_plugin.hpp"
-#include "cx_utils/LockSharedPtr.hpp"
+#include "cx_utils/lock_shared_ptr.hpp"
 
 #include <std_msgs/msg/empty.hpp>
