+++ cx_plugins/pddl_parser_plugin/src/pddl_parser_plugin.cpp
@@ -28,7 +28,7 @@
 
 #include "cx_pddl_parser_plugin/pddl_parser_plugin.hpp"
 #include "cx_plugin/clips_plugin.hpp"
-#include "cx_utils/LockSharedPtr.hpp"
+#include "cx_utils/lock_shared_ptr.hpp"
 
 #include <cx_utils/clips_env_context.hpp>
