+++ cx_plugins/protobuf_plugin/include/cx_protobuf_plugin/protobuf_plugin.hpp
@@ -30,7 +30,7 @@
 #include <string>
 
 #include "cx_plugin/clips_plugin.hpp"
-#include "cx_utils/LockSharedPtr.hpp"
+#include "cx_utils/lock_shared_ptr.hpp"
 
 namespace cx {
