+++ cx_plugins/ros_msgs_plugin/include/cx_ros_msgs_plugin/ros_msgs_plugin.hpp
@@ -6,7 +6,7 @@
 #include <string>
 
 #include "cx_plugin/clips_plugin.hpp"
-#include "cx_utils/LockSharedPtr.hpp"
+#include "cx_utils/lock_shared_ptr.hpp"
 #include <rosidl_typesupport_introspection_cpp/message_introspection.hpp>
 
 namespace cx {
