+++ cx_ros_comm_gen/templates/msg.jinja.hpp
@@ -27,7 +27,7 @@
 #include <string>
 
 #include "cx_plugin/clips_plugin.hpp"
-#include "cx_utils/LockSharedPtr.hpp"
+#include "cx_utils/lock_shared_ptr.hpp"
 
 #include "rclcpp/rclcpp.hpp"
 #include {{message_include_path}}
