+++ cx_utils/CMakeLists.txt
@@ -30,7 +30,6 @@ set(dependencies
 include_directories(include)
 
 set(CX_UTILS_SOURCES
-  src/LockSharedPtr.cpp
   src/map_skill.cpp
   src/clips_env_context.cpp
   src/param_utils.cpp
