--- cx_utils/src/LockSharedPtr.cpp
@@ -1,5 +1,0 @@
-// Licensed under GPLv2. See LICENSE file. Copyright Carologistics.
-
-#include "cx_utils/LockSharedPtr.hpp"
-
-namespace cx {}
